Character class version 1 working
I put some of the Start code into class to generalize a character
I gave characters Moves, in which the only move implemented so far is standing
There is one bug... the animation "sagatstand" doesn't load up until
after some time. I think it's because of a bug in ObjectTimeline
Hitboxes are cuurently implemented as a "struct" containing a reference to a Fixture
and containing some data concerning hitboxes, like damage and hitstun.
No hitboxes are up yet... I can't wait to test them -_-