Added projectile
Bugs include accepting input while in the move... that's a quick fix
Also, I can't get the projectiles to collide with the wall
Finally, there is a few frames where a part of Sagat's back leg doesn't have hitboxes
That could easily be fixed.
We have a solid projectile move. I did test it, and it subtracted health when hit.